requirements.txt
beautifulsoup4==4.7.1
requests==2.9.1
asyncio
discord.py
APScheduler==3.0.0
rq
urllib3
